-updated VS2017 WindowsTargetPlatformVersion from 10.0.17134.0 to 10.0.17763.0 -fixed up some incorrect project references in VS2017 files -fixed a couple random compiler warnings -updated resip/stack/test/Makefile.w32 to use newer link locations for OpenSSL -removed BOOST::shared_ptr and enabled_shared_from_this text from reTurn/README.txt

+++ resip/stack/ConnectionBase.cxx
@@ -1064,7 +1064,7 @@ ConnectionBase::getWriteBufferForExtraBytes(int currentPos, int extraBytes)
 {
    if (currentPos > 0 && extraBytes > 0)
    {
-      if ((currentPos + extraBytes) > mBufferSize)
+      if (((size_t)currentPos + (size_t)extraBytes) > mBufferSize)
       {
          mBufferSize = currentPos + extraBytes;
          char* buffer = MsgHeaderScanner::allocateBuffer((int)mBufferSize);
